Module structure for BilevelBenchmark.jl
