+++ Project.toml
@@ -1,17 +1,25 @@
-name = "BilevelBenchmark.jl"
+name = "BilevelBenchmark"
 uuid = "aa819f21-2bde-4658-8897-bab36330d9b7"
 authors = ["Valentin Dijon <vanadjy@gmail.com> and contributors"]
 version = "0.1.0"
 
 [deps]
+BilevelJuMP = "485130c0-026e-11ea-0f1a-6992cd14145c"
 Emporium = "4f286ecd-06d3-4492-bdef-0ca090b7d070"
 Formatting = "59287772-0a20-5a39-b81b-1366585eb4c0"
 GitHub = "bc5e4493-9b4d-5f90-b8aa-2b2bcaad7a26"
+Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
+JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 
 [compat]
+BilevelJuMP = "0.6.2"
 Emporium = "0.2.4"
 Formatting = "0.4"
 GitHub = "5"
+Ipopt = "1.10.6"
+JuMP = "1.26.0"
+LinearAlgebra = "1.11.0"
 julia = "1.6"
 
 [extras]
